Test for deleting invoice and some bugs fixed

+++ InvoiceTool.Application/UseCases/Invoices/DeleteInvoice.cs
@@ -7,6 +7,9 @@ public class DeleteInvoice(IInvoiceService invoiceService)
     private readonly IInvoiceService _invoiceService = invoiceService;
     public async Task<bool> Execute(int id)
     {
+        if (id <= 0)
+            throw new ArgumentException("Id cannot be zero or under.");
+
         var isDeleted = await _invoiceService.DeleteAsync(id);
 
         return isDeleted;
